refactor: extract the database provider catalogue out of compute-docker
The providers lived in `gfs-compute-docker`, in a module documented as
"Predefined container definitions". That was already the wrong home before
SQLite arrived: `compute-kubernetes` does not depend on this crate, yet the
CLI registers the same catalogue and then picks Docker *or* Kubernetes, so
the definitions were never Docker-specific. The proportions said so too —
the Docker adapter is 1,863 lines and the catalogue it carried is 5,081.

SQLite made it indefensible: a provider whose entire premise is that there
is no container, filed under container definitions, in a crate named for
Docker.

Move the catalogue to `gfs-db-providers`. The move is clean — the provider
files import only `gfs_domain` (plus rusqlite for the embedded engine) and
reference nothing from compute-docker. `compute-docker` is now just the
Docker compute adapter, and `rusqlite` is no longer among its dependencies.

Call sites move from `gfs_compute_docker::containers::register_all` to
`gfs_db_providers::register_all` — mechanical across cli, mcp, console and
the compute-docker integration tests.

No behaviour change. gfs-domain 255/255; gfs-db-providers 97/98, the one
failure being the pre-existing live-postgres test that needs a container
this machine cannot pull; e2e_sqlite 7/7; fmt clean.

This is the first of two steps. The provider still fabricates a placeholder
image and `compute_port: 0` to satisfy trait methods it never uses; the
trait split that makes that unrepresentable follows.
