+++ Cargo.toml
@@ -2,6 +2,7 @@
 resolver = "2"
 members = [
     "crates/adapters/compute-docker",
+    "crates/adapters/db-providers",
     "crates/adapters/compute-kubernetes",
     "crates/adapters/storage-apfs",
     "crates/adapters/storage-btrfs",
@@ -77,6 +78,7 @@ gfs-storage-btrfs          = { path = "crates/adapters/storage-btrfs" }
 gfs-storage-file           = { path = "crates/adapters/storage-file" }
 gfs-storage-kubernetes     = { path = "crates/adapters/storage-kubernetes" }
 gfs-compute-docker         = { path = "crates/adapters/compute-docker" }
+gfs-db-providers           = { path = "crates/adapters/db-providers" }
 gfs-compute-kubernetes     = { path = "crates/adapters/compute-kubernetes" }
 gfs-cli                    = { path = "crates/applications/cli" }
 gfs-mcp                    = { path = "crates/applications/mcp" }
