+++ crates/adapters/compute-docker/tests/query_in_instance_it.rs
@@ -7,7 +7,7 @@ use std::sync::Arc;
 use std::time::{Duration, Instant};
 
 use gfs_compute_docker::DockerCompute;
-use gfs_compute_docker::containers;
+use gfs_db_providers as containers;
 use gfs_domain::model::config::{EnvironmentConfig, GfsConfig, RuntimeConfig};
 use gfs_domain::ports::database_provider::InMemoryDatabaseProviderRegistry;
 use gfs_domain::usecases::repository::execute_query_usecase::ExecuteQueryUseCase;
