+++ crates/adapters/db-providers/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name                   = "gfs-db-providers"
+version.workspace      = true
+edition.workspace      = true
+authors.workspace      = true
+license.workspace      = true
+repository.workspace   = true
+rust-version.workspace = true
+
+[dependencies]
+gfs-domain = { workspace = true }
+rusqlite   = { workspace = true }
+tracing    = { workspace = true }
+
+[dev-dependencies]
+serde_json = { workspace = true }
+tempfile   = { workspace = true }
