+++ crates/adapters/db-providers/src/detach_remote.sql
