+++ crates/applications/cli/Cargo.toml
@@ -35,6 +35,7 @@ path = "tests/schema/extract_clickhouse.rs"
 gfs-domain     = { workspace = true }
 gfs-telemetry  = { workspace = true }
 gfs-compute-docker        = { workspace = true }
+gfs-db-providers   = { workspace = true }
 gfs-compute-kubernetes    = { workspace = true }
 gfs-storage-file   = { workspace = true }
 gfs-storage-kubernetes = { workspace = true }
