+++ crates/applications/cli/src/commands/cmd_checkout.rs
@@ -46,7 +46,7 @@ pub async fn checkout(
     let repository: Arc<dyn Repository> = Arc::new(GfsRepository::new());
     let compute: Arc<dyn Compute> = compute_for_repo(&repository, &repo_path).await?;
     let registry = Arc::new(InMemoryDatabaseProviderRegistry::new());
-    gfs_compute_docker::containers::register_all(registry.as_ref())
+    gfs_db_providers::register_all(registry.as_ref())
         .context("failed to register database providers")?;
 
     // Kubernetes runtime needs PVC restore; the generic checkout use case assumes filesystem snapshots.
