+++ crates/applications/cli/src/commands/cmd_export.rs
@@ -33,7 +33,7 @@ pub async fn run(
     let _ = id; // container name override is reserved for future use; use case reads from config.
 
     let registry = Arc::new(InMemoryDatabaseProviderRegistry::new());
-    gfs_compute_docker::containers::register_all(registry.as_ref())
+    gfs_db_providers::register_all(registry.as_ref())
         .context("failed to register database providers")?;
 
     // A lazy clone only holds rows for tables that have actually been read, and
