+++ crates/applications/cli/src/commands/cmd_import.rs
@@ -33,7 +33,7 @@ pub async fn run(
     let _ = id; // container name override reserved for future use.
 
     let registry = Arc::new(InMemoryDatabaseProviderRegistry::new());
-    gfs_compute_docker::containers::register_all(registry.as_ref())
+    gfs_db_providers::register_all(registry.as_ref())
         .context("failed to register database providers")?;
 
     let format_str = format.as_deref().unwrap_or("");
