+++ crates/applications/cli/src/commands/cmd_schema.rs
@@ -42,7 +42,7 @@ pub async fn run_extract(
     let compute = compute_for_repo(&repository, &repo_path).await?;
 
     let registry = Arc::new(InMemoryDatabaseProviderRegistry::new());
-    gfs_compute_docker::containers::register_all(registry.as_ref())
+    gfs_db_providers::register_all(registry.as_ref())
         .context("failed to register database providers")?;
 
     let use_case = ExtractSchemaUseCase::new(compute, registry);
