+++ crates/applications/mcp/src/tools.rs
@@ -4,8 +4,8 @@ use std::path::PathBuf;
 use std::sync::Arc;
 
 use gfs_compute_docker::DockerCompute;
-use gfs_compute_docker::containers;
 use gfs_compute_kubernetes::KubernetesCompute;
+use gfs_db_providers as containers;
 use gfs_domain::adapters::gfs_repository::GfsRepository;
 use gfs_domain::model::config::{GfsConfig, RuntimeConfig};
 use gfs_domain::model::datasource::diff::compute_schema_diff;
